Moved back to Primus/WS
I had a think about it, and we want the one-to-many information about clients connected to the server.  Specifically, we easily want to know what to distribute to what slave, and we really really want to know what slaves are online without having to build some sort of disastrous heartbeat mechanism over AMQP.  Just no.